requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flash==1.0.3
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
Werkzeug==3.1.3
